+++ test/fixtures/index.ts
@@ -1062,7 +1062,9 @@ export const fspiopIso20022Sources = {
             Ccy: 'XXY',
             ActiveCurrencyAndAmount: '23.55'
           },
-          ChrgBr: 'DEBT'
+          InstrForCdtrAgt: {
+            InstrInf: 'RECEIVE'
+          }
         }
       }
     },
@@ -1132,7 +1134,9 @@ export const fspiopIso20022Sources = {
             Ccy: 'XXY',
             ActiveCurrencyAndAmount: '23.55'
           },
-          ChrgBr: 'DEBT'
+          InstrForCdtrAgt: {
+            InstrInf: 'RECEIVE'
+          }
         }
       }
     },
@@ -1817,7 +1821,9 @@ export const expectedFspiopIso20022Targets = (target: GenericObject) => ({
             Ccy: 'XXY',
             ActiveCurrencyAndAmount: '23.55'
           },
-          ChrgBr: 'DEBT'
+          InstrForCdtrAgt: {
+            InstrInf: 'RECEIVE'
+          }
         }
       }
     },
@@ -1887,7 +1893,9 @@ export const expectedFspiopIso20022Targets = (target: GenericObject) => ({
             Ccy: 'XXY',
             ActiveCurrencyAndAmount: '23.55'
           },
-          ChrgBr: 'DEBT'
+          InstrForCdtrAgt: {
+            InstrInf: 'RECEIVE'
+          }
         },
         GrpHdr: {
           MsgId: getProp(target, 'body.GrpHdr.MsgId'),