+++ test/unit/facades/fspiop.test.ts
@@ -408,20 +408,6 @@ describe('FSPIOPTransformFacade tests', () => {
       test('should transform POST FX quotes payload from FSPIOP to FSPIOP ISO 20022', async () => {
         await testCase(fspiopSources.fxQuotes.post, FspiopTransformFacade.fxQuotes.post, expected('fxQuotes.post'))();
       });
-      test('should not apply mutation on target if override mapping is set', async () => {
-        const source = { ...fspiopSources.fxQuotes.post };
-        const overrideMapping = FSPIO20022PMappings.fxQuotes_reverse.post;
-        const target = await FspiopTransformFacade.fxQuotes.post(source, { overrideMapping });
-        expect(target).toHaveProperty('body');
-        expect(getProp(target, 'body.CdtTrfTxInf.ChrgBr')).toBeUndefined();
-      });
-      test('should transform POST FX quotes payload from FSPIOP to FSPIOP ISO 20022 with amountType === SEND', async () => {
-        const source = { ...fspiopSources.fxQuotes.post };
-        setProp(source, 'body.conversionTerms.amountType', 'SEND');
-        const target = await FspiopTransformFacade.fxQuotes.post(source);
-        expect(target).toHaveProperty('body');
-        expect(getProp(target, 'body.CdtTrfTxInf.ChrgBr')).toBe('CRED');
-      });
     });
     describe('PUT /fxQuotes', () => {
       test('should throw if source is wrongly typed', async () => {
@@ -432,20 +418,6 @@ describe('FSPIOPTransformFacade tests', () => {
       });
       test('should transform PUT FX quotes payload from FSPIOP to FSPIOP ISO 20022', async () => {
         await testCase(fspiopSources.fxQuotes.put, FspiopTransformFacade.fxQuotes.put, expected('fxQuotes.put'))();
-      });
-      test('should not apply mutation on target if override mapping is set', async () => {
-        const source = { ...fspiopSources.fxQuotes.put };
-        const overrideMapping = FSPIO20022PMappings.fxQuotes_reverse.put;
-        const target = await FspiopTransformFacade.fxQuotes.put(source, { overrideMapping });
-        expect(target).toHaveProperty('body');
-        expect(getProp(target, 'body.CdtTrfTxInf.ChrgBr')).toBeUndefined();
-      });
-      test('should transform PUT FX quotes payload from FSPIOP to FSPIOP ISO 20022 with amountType === SEND', async () => {
-        const source = { ...fspiopSources.fxQuotes.put };
-        setProp(source, 'body.conversionTerms.amountType', 'SEND');
-        const target = await FspiopTransformFacade.fxQuotes.put(source);
-        expect(target).toHaveProperty('body');
-        expect(getProp(target, 'body.CdtTrfTxInf.ChrgBr')).toBe('CRED');
       });
     });
     describe('PUT /fxQuotes/{ID}/error', () => {